kv_service/requirements.txt
djangorestframework==3.16.0
cryptography==45.0.4
psycopg2-binary==2.9.10
